pyproject.toml
[project]
name = "sales-forecast"
version = "0.1.0"
description = "End-to-end Vertex AI forecasting pipeline"
authors = [
    {name = "dev.thryvma@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.7.2,<2.0.0)", 
    "lightgbm (>=4.6.0,<5.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "google-cloud-bigquery (>=3.38.0,<4.0.0)",
    "google-cloud-storage (>=2.10.0,<3.0.0)",
    "jupyterlab (>=4.4.10,<5.0.0)",
    "ipykernel (>=7.0.1,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "db-dtypes (>=1.4.3,<2.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "kfp (>=2.7.0,<3.0.0)", 
    "google-cloud-aiplatform (>=1.56.0,<2.0.0)" 
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
